Rename: "category -> categories, loaderMap -> categoryLoader"
내용과 의미를 더 명확하게 전달하기위해 변경
