+++ static/js/view/categories/home.js
@@ -1,4 +1,4 @@
-import { renderer } from '../../renderer.js';
+import { renderer } from '../renderer.js';
 
 const home = () => {
   renderer.preparing();
