+++ static/js/view/categories/webnovel.js
@@ -1,4 +1,4 @@
-import { renderer } from '../../renderer.js';
+import { renderer } from '../renderer.js';
 
 const webnovel = {
   home: (DATA) => {},
