+++ static/js/view/categories/webtoon.js
@@ -1,4 +1,4 @@
-import { renderer } from '../../renderer.js';
+import { renderer } from '../renderer.js';
 
 const webtoon = {
   home: (DATA) => {
